RemovedCommandBase
Changed SwerveBase Code from CommandBase to State Machine
